refactor(cli): move snapshot isolation to kilocode/
